#99, made makeDefaultHttpClientWithCVurlProxySelector private and fixed test. Moved ProxySelectorMock to separate class
