+++ src/test/java/coresearch/cvurl/io/request/ProxySelectorMock.java
@@ -1,0 +1,26 @@
+package coresearch.cvurl.io.request;
+
+import java.io.IOException;
+import java.net.Proxy;
+import java.net.ProxySelector;
+import java.net.SocketAddress;
+import java.net.URI;
+import java.util.List;
+
+final class ProxySelectorMock extends ProxySelector {
+    private final List<Proxy> expectedProxies;
+
+    public ProxySelectorMock(List<Proxy> expectedProxies) {
+        this.expectedProxies = expectedProxies;
+    }
+
+    @Override
+    public List<Proxy> select(URI uri) {
+        return expectedProxies;
+    }
+
+    @Override
+    public void connectFailed(URI uri, SocketAddress sa, IOException ioe) {
+        // do nothing
+    }
+}
